add one extra layer to protect data

+++ src/components/Dashbord.jsx
@@ -1,5 +1,11 @@
 import {
   Alert,
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
   AlertIcon,
   Box,
   Button,
@@ -8,7 +14,7 @@ import {
   GridItem,
   Image,
 } from "@chakra-ui/react";
-import { useContext, useEffect, useState } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import { AuthContext } from "../Context/AuthContext";
 import { Navigate } from "react-router-dom";
 import axios from "axios";
@@ -24,12 +30,17 @@ export const Dashbord = () => {
 
   const [isLoading, setIsLoading] = useState(false);
 
+  const [isOpen, setIsOpen] = useState(false);
+  const onClose = () => setIsOpen(false);
+  const cancelRef = useRef();
+
   useEffect(() => {
     getDetails();
   }, []);
 
   const handleDelete = (id) => {
     // console.log(id._id);
+    onClose();
     setDeleteProgress(true);
     axios
       .delete(`https://kapoorkartik.herokuapp.com/kaizenapp/${id._id}`)
@@ -127,17 +138,49 @@ export const Dashbord = () => {
                     color="red"
                   />
                 ) : (
-                  <Button
-                    size="md"
-                    height="30px"
-                    width="120px"
-                    border="1px"
-                    bg="red"
-                    ml="30%"
-                    onClick={() => handleDelete(e)}
-                  >
-                    Delete
-                  </Button>
+                  <>
+                    <Button
+                      size="md"
+                      height="30px"
+                      width="120px"
+                      border="1px"
+                      bg="red"
+                      ml="30%"
+                      onClick={() => setIsOpen(true)}
+                    >
+                      Delete
+                    </Button>
+                    <AlertDialog
+                      isOpen={isOpen}
+                      leastDestructiveRef={cancelRef}
+                      onClose={onClose}
+                    >
+                      <AlertDialogOverlay>
+                        <AlertDialogContent>
+                          <AlertDialogHeader fontSize="lg" fontWeight="bold">
+                            Delete
+                          </AlertDialogHeader>
+
+                          <AlertDialogBody>
+                            Are you sure? You can't undo this action afterwards.
+                          </AlertDialogBody>
+
+                          <AlertDialogFooter>
+                            <Button ref={cancelRef} onClick={onClose}>
+                              Cancel
+                            </Button>
+                            <Button
+                              colorScheme="red"
+                              onClick={() => handleDelete(e)}
+                              ml={3}
+                            >
+                              Delete
+                            </Button>
+                          </AlertDialogFooter>
+                        </AlertDialogContent>
+                      </AlertDialogOverlay>
+                    </AlertDialog>
+                  </>
                 )}
               </GridItem>
             </Grid>
@@ -147,3 +190,45 @@ export const Dashbord = () => {
     </>
   );
 };
+/**
+ * function AlertDialogExample() {
+  const [isOpen, setIsOpen] = React.useState(false)
+  const onClose = () => setIsOpen(false)
+  const cancelRef = React.useRef()
+
+  return (
+    <>
+      <Button colorScheme='red' onClick={() => setIsOpen(true)}>
+        Delete Customer
+      </Button>
+
+      <AlertDialog
+        isOpen={isOpen}
+        leastDestructiveRef={cancelRef}
+        onClose={onClose}
+      >
+        <AlertDialogOverlay>
+          <AlertDialogContent>
+            <AlertDialogHeader fontSize='lg' fontWeight='bold'>
+              Delete Customer
+            </AlertDialogHeader>
+
+            <AlertDialogBody>
+              Are you sure? You can't undo this action afterwards.
+            </AlertDialogBody>
+
+            <AlertDialogFooter>
+              <Button ref={cancelRef} onClick={onClose}>
+                Cancel
+              </Button>
+              <Button colorScheme='red' onClick={onClose} ml={3}>
+                Delete
+              </Button>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialogOverlay>
+      </AlertDialog>
+    </>
+  )
+}
+ */
